Change some integer type i32 to c_int

+++ src/syscall/lseek.rs
@@ -1,3 +1,5 @@
+use std::ffi::c_int;
+
 use libc::{SEEK_CUR, SEEK_DATA, SEEK_END, SEEK_HOLE, SEEK_SET};
 
 use crate::syscall::common::*;
@@ -12,15 +14,15 @@ const LSEEK_WHENCE_DESCS: &[FlagDesc] = &[
 
 #[repr(C)]
 struct LseekArgs {
-    fd: i32,
+    fd: c_int,
     offset: libc::off_t,
-    whence: i32,
+    whence: c_int,
 }
 unsafe impl plain::Plain for LseekArgs {}
 
 pub(super) fn handle_lseek_args(args: &[u8]) -> String {
     let lseek = get_args::<LseekArgs>(args);
-    let whence = format_flags(lseek.whence as u32, '|', &LSEEK_WHENCE_DESCS);
+    let whence = format_flags(lseek.whence, '|', &LSEEK_WHENCE_DESCS);
 
     return format!("{}, {}, {}", lseek.fd, lseek.offset, whence);
 }
